requirements.txt
flask
pandas
pillow
open_clip_torch
torch
tqdm
jsonify